% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{adapt_image}
\alias{adapt_image}
\title{Fits an image to the given size.}
\usage{
adapt_image(input, output = NULL, width = 800, height = 600)
}
\arguments{
\item{input}{A string, image file.}

\item{output}{A string, image file.}

\item{width}{A integer, width of the image.}

\item{height}{A integer, height of the image.}
}
\value{
A string, new image file.
}
\description{
Fits an image to the given size.
}
\examples{
#

}
\concept{common functions}
