% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{adapt_image_folder}
\alias{adapt_image_folder}
\title{Fits an image to the given size.}
\usage{
adapt_image_folder(input, output = NULL, width = 800, height = 600)
}
\arguments{
\item{input}{A string, image folder.}

\item{output}{A string, image folder.}

\item{width}{A integer, width of the image.}

\item{height}{A integer, height of the image.}
}
\value{
A string, output image folder.
}
\description{
Fits an image to the given size.
}
\examples{
#

}
\seealso{
Other common functions: 
\code{\link{adapt_image}()}
}
\concept{common functions}
