% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_raster.R
\name{aggregate_raster}
\alias{aggregate_raster}
\title{Aggregate rasters in a folder}
\usage{
aggregate_raster(dir, out_dir, factor = 100)
}
\arguments{
\item{dir}{A string or string vector.}

\item{out_dir}{A string, output folder.}

\item{factor}{A integer, aggregation factor.}
}
\value{
A string vector.
}
\description{
Aggregate rasters in a folder
}
