% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_vector.R
\name{clip_vector}
\alias{clip_vector}
\title{Clip a vector layer based on a polygon}
\usage{
clip_vector(vector, polygon)
}
\arguments{
\item{vector}{A `sf` vector layer.}

\item{polygon}{A `sf` polygon layer.}
}
\value{
A `sf` vector layer.
}
\description{
We must ensure that the layer to be cut has the correct encoding.
}
\details{
Project the result to the CRS of the clipping polygon.
}
\examples{
#

}
\seealso{
Other transformation functions: 
\code{\link{clip_multipoligon}()},
\code{\link{clip_raster}()},
\code{\link{copy_styles_layer}()},
\code{\link{copy_styles_layer_names}()},
\code{\link{get_layer_categories}()}
}
\concept{transformation functions}
