% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_raster.R
\name{compose_raster}
\alias{compose_raster}
\title{Compose a raster layer from multiple files}
\usage{
compose_raster(dir, out_file = NULL)
}
\arguments{
\item{dir}{A string or string vector, folder names.}

\item{out_file}{A string, output file name (without extension).}
}
\value{
A `terra` raster layer.
}
\description{
The folder name or vector of folder names that contain the files is indicated.
}
\details{
We can also indicate the name (with the path and without the extension) of the
virtual raster file; if none is indicated, a temporary one is used.
}
\examples{
#

}
\concept{generation functions}
