% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_styles_layer.R
\name{copy_styles_layer_names}
\alias{copy_styles_layer_names}
\title{Copy styles layer from one source `GeoPackage` to a PostGIS database.}
\usage{
copy_styles_layer_names(from, to, layers)
}
\arguments{
\item{from}{A GeoPackage file name.}

\item{to}{A database connection.}

\item{layers}{A vector of layer names.}
}
\value{
`obj`, invisibly.
}
\description{
Assigns the first style from the source the given layers in the destination.
}
\examples{
#

}
\seealso{
Other transformation functions: 
\code{\link{clip_multipoligon}()},
\code{\link{clip_raster}()},
\code{\link{clip_vector}()},
\code{\link{copy_styles_layer}()}
}
\concept{transformation functions}
