% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bbox.R
\name{generate_bbox}
\alias{generate_bbox}
\title{Generate a Bounding Box as an sf Object}
\usage{
generate_bbox(layer)
}
\arguments{
\item{layer}{An `sf` object, representing the input vector layer.}
}
\value{
An `sf` object representing the bounding box of the input layer.
}
\description{
This function takes an `sf` object as input and returns a new `sf` object
representing the bounding box (minimum bounding rectangle) of the input layer.
}
\examples{
source_gpkg <- system.file("extdata/sigugr.gpkg", package = "sigugr")
lanjaron <- sf::st_read(source_gpkg, layer = "lanjaron", quiet = TRUE)

bbox_layer <- generate_bbox(lanjaron)

}
