% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rasters.R
\name{get_nexus}
\alias{get_nexus}
\title{Get nexus}
\usage{
get_nexus(name)
}
\arguments{
\item{name}{A string representing a folder path.}
}
\value{
A string. If the input ends with "/", it returns an empty string.
  Otherwise, it returns "/".
}
\description{
Determines the appropriate path separator based on whether a folder path
ends with a trailing slash or not.
}
\keyword{internal}
