% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_publish_layer.R
\name{gs_publish_layer}
\alias{gs_publish_layer}
\title{Publish a layer to GeoServer}
\usage{
gs_publish_layer(
  geoserver_url,
  user,
  password,
  workspace,
  data_store,
  layer_name,
  title = NULL
)
}
\arguments{
\item{geoserver_url}{A string, the base URL of the GeoServer instance.}

\item{user}{A string, GeoServer username.}

\item{password}{A string, GeoServer password.}

\item{workspace}{A string, the name of the GeoServer workspace.}

\item{data_store}{A string, the name of the data store where the layer resides.}

\item{layer_name}{A string, the name of the layer to publish.}

\item{title}{A string, an optional title for the layer. Defaults to the layer
name if not provided.}
}
\value{
Integer. Returns `0` on success or `1` on failure. Prints a message
  describing the result.
}
\description{
Publishes a vector layer to GeoServer using the REST API. Returns a status code
and prints an appropriate message indicating success or failure.
}
\examples{
\dontrun{
gs_publish_layer(
  geoserver_url = "http://my-geoserver.com/geoserver",
  user = "admin",
  password = "mypassword",
  workspace = "my_workspace",
  data_store = "my_datastore",
  layer_name = "my_layer",
  title = "My Layer Title"
)
}
}
