% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_publish_raster.R
\name{gs_publish_raster}
\alias{gs_publish_raster}
\title{Publish a Raster to GeoServer}
\usage{
gs_publish_raster(geoserver_url, user, password, workspace, data_store, raster)
}
\arguments{
\item{geoserver_url}{A character string specifying the base URL of the GeoServer instance (e.g., "http://localhost:8080/geoserver").}

\item{user}{A character string for the GeoServer username with publish permissions.}

\item{password}{A character string for the GeoServer password corresponding to the specified user.}

\item{workspace}{A character string specifying the target workspace where the raster will be published.}

\item{data_store}{A character string specifying the name of the target data store in GeoServer.}

\item{raster}{A character string specifying the file path to the GeoTIFF raster file to be uploaded.}
}
\value{
An integer indicating the status of the operation:
  - `0`: The raster was published successfully.
  - `1`: An error occurred during the publishing process.
}
\description{
This function publishes a GeoTIFF raster file to a specified workspace and data store on a GeoServer instance.
}
\details{
The function uses the GeoServer REST API to upload a GeoTIFF raster file to a specified workspace and data store.
The raster is sent using an HTTP `PUT` request to the GeoServer endpoint, and the response is checked for success.
A message is printed to indicate whether the operation succeeded or failed, and any error messages from GeoServer are displayed.
}
\examples{
\dontrun{
# Publish a raster to GeoServer
geoserver_url <- "http://localhost:8080/geoserver"
user <- "admin"
password <- "geoserver"
workspace <- "example_workspace"
data_store <- "example_datastore"
raster <- "path/to/raster.tif"

gs_publish_raster(
  geoserver_url = geoserver_url,
  user = user,
  password = password,
  workspace = workspace,
  data_store = data_store,
  raster = raster
)
}
}
\seealso{
Other publish to GeoServer: 
\code{\link{gs_publish_bands}()},
\code{\link{gs_publish_layer}()}
}
\concept{publish to GeoServer}
