% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{list_dir_rasters}
\alias{list_dir_rasters}
\title{List Raster Files in a Directory}
\usage{
list_dir_rasters(dir)
}
\arguments{
\item{dir}{A character string specifying the path to the directory to search.}
}
\value{
A character vector containing the full paths of all matching files.
If no matching files are found, an empty character vector is returned.
}
\description{
This function scans a directory and recursively lists all raster files
with extensions `.tif` or `.jp2`. The search is case-insensitive.
}
\keyword{internal}
