% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_write_bands.R
\name{pg_write_bands}
\alias{pg_write_bands}
\title{Save raster bands in separate tables in PostGIS}
\usage{
pg_write_bands(sr, conn, prefix = NULL, postfix = NULL)
}
\arguments{
\item{sr}{A `SpatRaster` object.}

\item{conn}{A connection to a PostGIS database.}

\item{prefix}{A string, table name prefix to the band name.}

\item{postfix}{A string, table name postfix to the band name.}
}
\value{
A `SpatRaster` object.
}
\description{
Save raster bands in separate tables in PostGIS
}
\examples{
#

}
\seealso{
Other save functions: 
\code{\link{pg_write_satres_bands}()}
}
\concept{save functions}
