% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_write_raster_file_bands.R
\name{pg_write_raster_file_bands}
\alias{pg_write_raster_file_bands}
\title{Save raster file bands in separate tables in PostGIS}
\usage{
pg_write_raster_file_bands(file, conn, prefix = NULL, postfix = NULL)
}
\arguments{
\item{file}{A string vector, raster file names.}

\item{conn}{A connection to a PostGIS database.}

\item{prefix}{A string, table name prefix to the band name.}

\item{postfix}{A string, table name postfix to the band name.}
}
\value{
A string vector, raster file names.
}
\description{
Save raster file bands in separate tables in PostGIS
}
\examples{
#

}
