% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_write_satres_bands.R
\name{pg_write_satres_bands}
\alias{pg_write_satres_bands}
\title{Save `satres` bands in separate tables in PostGIS}
\usage{
pg_write_satres_bands(sr, conn, prefix = NULL, postfix = NULL)
}
\arguments{
\item{sr}{A `satres` object.}

\item{conn}{A connection to a PostGIS database.}

\item{prefix}{A string, table name prefix to the resolution names.}

\item{postfix}{A string, table name postfix to the resolution names.}
}
\value{
A `satres` object.
}
\description{
Save `satres` bands in separate tables in PostGIS
}
\examples{
#

}
