% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_publish_layer.R
\name{publish_layer}
\alias{publish_layer}
\alias{publish_layer.geoserver}
\title{Publish a Vector Layer to GeoServer}
\usage{
publish_layer(gso, layer, title)

\method{publish_layer}{geoserver}(gso, layer, title = NULL)
}
\arguments{
\item{gso}{An object of class `geoserver` containing GeoServer connection details.}

\item{layer}{A string, the name of the layer to publish.}

\item{title}{A string, an optional title for the layer. Defaults to the layer
name if not provided.}
}
\value{
An object of class `geoserver` or NULL if an error occurred.
}
\description{
Publishes a vector layer to GeoServer using the REST API.
}
\details{
Prints an appropriate message indicating success or failure.
}
\examples{
\dontrun{
}
}
\seealso{
Other publish to GeoServer: 
\code{\link{geoserver}()},
\code{\link{gs_publish_bands}()},
\code{\link{gs_publish_raster}()},
\code{\link{register_datastore_geopackage}()},
\code{\link{register_datastore_postgis}()}
}
\concept{publish to GeoServer}
