% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_register_tiff.R
\name{register_datastore_tiff}
\alias{register_datastore_tiff}
\alias{register_datastore_tiff.geoserver}
\title{Register a TIFF as a DataStore in GeoServer}
\usage{
register_datastore_tiff(gso, datastore, tiff)

\method{register_datastore_tiff}{geoserver}(gso, datastore, tiff = NULL)
}
\arguments{
\item{datastore}{A character string. The name of the datastore to be created.}

\item{tiff}{A character string specifying the full file path to the TIFF
file. The default is `NULL`, which assumes the datastore has already been registered.}

\item{An}{object of class `geoserver`.}
}
\value{
An object of class `geoserver` or NULL if an error occurred.
}
\description{
This function registers a TIFF as a `datastore` in a specified GeoServer workspace
using the REST API.
}
\details{
If the `datastore` has already been registered previously, there is no need to specify the
TIFF. For subsequent operations, that `datastore` will be used.

In any case, prints an appropriate message.
}
\examples{
\dontrun{
gso <- geoserver(
  url = "http://localhost:8080/geoserver",
  user = "admin",
  password = "geoserver",
  workspace = "sigugr_test"
)

source_tif <- system.file("extdata/sat.tif", package = "sigugr")

gso <- gso |>
  register_datastore_tiff("sigugr-tiff", source_tif)
}

}
\seealso{
Other publish to GeoServer: 
\code{\link{geoserver}()},
\code{\link{gs_publish_bands}()},
\code{\link{gs_publish_raster}()},
\code{\link{publish_layer}()},
\code{\link{register_datastore_geopackage}()},
\code{\link{register_datastore_postgis}()}
}
\concept{publish to GeoServer}
