% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres_minmax.R
\name{satres_minmax}
\alias{satres_minmax}
\title{Get minmax}
\usage{
satres_minmax(sr)
}
\arguments{
\item{sr}{A `satres` object.}
}
\value{
A string vector.
}
\description{
Get all minmax values in vector form to check them.
}
