% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_raster.R
\name{store_raster}
\alias{store_raster}
\title{Store Raster to PostGIS}
\usage{
store_raster(sr, conn, schema = "public", table_name)
}
\arguments{
\item{sr}{A `terra::SpatRaster` object containing the raster bands to write.}

\item{conn}{A database connection object to a PostGIS database (e.g., from `RPostgres::dbConnect`).}

\item{schema}{A string specifying the schema in the PostGIS database where
the raster layers will be stored. Default is `"public"`.}

\item{table_name}{A string, table name.}
}
\value{
Invisibly returns a character vector of the names of the tables written to PostGIS.
}
\description{
Stores all bands of a raster (`terra::SpatRaster`) to a specified schema in a
PostGIS database. All bands are written in the same table in the database.
}
\details{
Transforms the table name according to the Snake Case convention.
}
\examples{
\dontrun{
conn <- DBI::dbConnect(
  RPostgres::Postgres(),
  dbname = "mydb",
  host = "localhost",
  user = "user",
  password = "password"
)

sr <- terra::rast(nrows = 10, ncols = 10, nlyrs = 3, vals = runif(300))

tables <- store_raster(sr, conn, schema = "geodata", table_name = "example_raster")

DBI::dbDisconnect(conn)
}

}
\seealso{
Other write to PostGIS: 
\code{\link{store_bands}()},
\code{\link{store_layers}()}
}
\concept{write to PostGIS}
