% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_multipoligon.R
\name{clip_multipoligon}
\alias{clip_multipoligon}
\title{Clip a multipoligon vector layer based on a polygon}
\usage{
clip_multipoligon(vector, polygon)
}
\arguments{
\item{vector}{A `sf` multipoligon vector layer.}

\item{polygon}{A `sf` polygon layer.}
}
\value{
A `sf` vector layer.
}
\description{
We will use this function only if function `clip_vector()` gives an error.
The error I have found is `ParseException: Unknown WKB type 12`, it has to do
with type *MULTIPOLYGON*.
}
\details{
The solution adopted is based on the one described in
https://gis.stackexchange.com/questions/389814/r-st-centroid-geos-error-unknown-wkb-type-12

We must ensure that the layer to be cut has the correct encoding.

Project the result to the CRS of the clipping polygon.
}
\examples{
#

}
\seealso{
Other transformation functions: 
\code{\link{clip_raster}()},
\code{\link{clip_vector}()},
\code{\link{copy_styles_layer}()},
\code{\link{copy_styles_layer_names}()}
}
\concept{transformation functions}
