% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_raster.R
\name{clip_raster}
\alias{clip_raster}
\title{Clip a raster based on a polygon}
\usage{
clip_raster(raster, polygon)
}
\arguments{
\item{raster}{A `terra` raster to be clipped.}

\item{polygon}{A `sf` polygon layer used for clipping.}
}
\value{
A `terra` raster clipped to the extent of the polygon.
}
\description{
This function clips a raster using a polygon, preserving the coordinate
reference system (CRS) of the raster.
}
\examples{
source_gpkg <- system.file("extdata", "sigugr.gpkg", package = "sigugr")
p <-sf::st_read(source_gpkg, layer = 'lanjaron', quiet = TRUE)

source_tif <- system.file("extdata", "sat.tif", package = "sigugr")
r <- terra::rast(source_tif)

result <- clip_raster(r, p)

}
