% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_raster.R
\name{clip_raster}
\alias{clip_raster}
\title{Clip a raster layer based on a polygon}
\usage{
clip_raster(raster, polygon)
}
\arguments{
\item{raster}{A `terra` raster layer to be clipped.}

\item{polygon}{A `sf` polygon layer used for clipping.}
}
\value{
A `terra` raster layer clipped to the extent of the polygon.
}
\description{
This function clips a raster layer using a polygon, preserving the coordinate reference system (CRS) of the raster.
}
\details{
The function first transforms the polygon to match the CRS of the raster, then crops the raster to the extent of the polygon.
Finally, it applies a mask to retain only the areas within the polygon.
}
\examples{
#

}
