% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigugr_styles.R
\name{copy_styles_layer}
\alias{copy_styles_layer}
\title{Copy Layer Styles from Source to Destination}
\usage{
copy_styles_layer(from, to)
}
\arguments{
\item{from}{A data origin. This can be:
- A string representing the path to a GeoPackage file.
- A `DBI` database connection object to a PostGIS database, created using [RPostgres::dbConnect()].}

\item{to}{A string representing the path to the destination GeoPackage file.}
}
\value{
The updated `layer_styles` table, returned invisibly.
}
\description{
Copies the first style definition from a source (either a GeoPackage file or a
PostGIS connection) and assigns it to all layers in the destination GeoPackage.
}
\details{
The function reads the first style from the `layer_styles` table in the source
GeoPackage or PostGIS database. This style is then applied to all layers in the
destination GeoPackage.
}
\examples{
\dontrun{
source_gpkg <- "source.gpkg"
dest_gpkg <- "destination.gpkg"

copy_styles_layer(from = source_gpkg, to = dest_gpkg)
}
}
\seealso{
\code{\link{copy_styles_layer_names}}

Other independent functions: 
\code{\link{clip_multipoligon}()},
\code{\link{clip_vector}()},
\code{\link{copy_styles_layer_names}()},
\code{\link{get_layer_categories}()}
}
\concept{independent functions}
