% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_styles_layer.R
\name{get_layer_categories}
\alias{get_layer_categories}
\title{Get layer categories from a styles layer from one source `GeoPackage`}
\usage{
get_layer_categories(from, r_clc)
}
\arguments{
\item{from}{A GeoPackage file name.}

\item{r_clc}{A `terra` raster.}
}
\value{
`categories` A data frame of categories.
}
\description{
It obtains them from the first defined style.
}
\examples{
#

}
\seealso{
Other transformation functions: 
\code{\link{clip_multipoligon}()},
\code{\link{clip_raster}()},
\code{\link{clip_vector}()},
\code{\link{copy_styles_layer}()},
\code{\link{copy_styles_layer_names}()}
}
\concept{transformation functions}
